fix: move card component
